requirements.txt
requests>=2.31,<3
beautifulsoup4>=4.12,<5
openai>=1.30,<2
resend>=2.0,<3
python-dotenv>=1.0,<2
